+++ CHANGES.md
@@ -14,6 +14,7 @@
 | Removed `if (f > 5)` in `MainLoop()` in `AsyncLogger` | It stopped any Log from making more than five logs. |
 | Changed `_lines` in `AsyncLogger` from a `List<LogLine>` into `ConcurrectQueue<LogLine>`. This also caused several other changes to take this type-change into account | Since the list was being accessed by two threads, the one adding to it in `Program.cs` and the `MainLoop()`, it needed to be able to handle several things trying to access it at once (thread-safe) |
 | Created test for `AsyncLogger` | |
+| Turned all instances of the `DateTime`-data type into `DateTimeOffset` | In order to use the `TimeProvider` normally and `FakeTimeProvider` for testing. |
 
 ## Changes I wish to make
 
@@ -25,3 +26,4 @@
 - Make test to prove that stop behavior works as intended.
 - Consider making locks for `_exit` and `_QuitWithFlush` in `AsyncLogger` since they can get accessed by more than one thread at the same time.
 - Consider renaming `MainLoop()` in `AsyncLogger` to something that explains what it does.
+- Consider using `DateTimeOff.GetUtcNow()` instead of `DateTime.GetLocalNow()`.